Fix: [ALAS] Handle another format of multiscreen warning (#5682)
